backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
cachetools==6.2.4
celery==5.5.3
certifi==2025.7.14
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-filter==25.1
django-redis==6.0.0
django-rest-framework==0.1.0
django-unfold==0.63.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
exceptiongroup==1.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.60.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
loguru==0.7.3
msgpack==1.1.2
oauthlib==3.3.1
packaging==25.0
phonenumbers==9.0.10
pluggy==1.6.0
prompt_toolkit==3.0.51
proto-plus==1.27.0
protobuf==5.29.5
psycopg2==2.9.11
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.3.1
pytest==8.4.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tomli==2.2.1
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
win32_setctime==1.2.0
zope.interface==8.2
